requirements.txt
mysql-connector-python==9.3.0
python-dotenv==1.1.0
tk==0.1.0
watchdog==6.0.0
